fixed spearman rank to handle large number of attributes fixed tanimoto to return values between 0 and 1
git-svn-id: https://svn.code.sf.net/p/java-ml/code/trunk@862 3162d2cb-3e1f-0410-96b0-914c58786302

+++ net/sf/javaml/distance/SpearmanRankCorrelation.java
@@ -23,8 +23,8 @@ public class SpearmanRankCorrelation extends AbstractCorrelation {
     public double calculateDistance(Instance a, Instance b) {
         if (a.size() != b.size())
             throw new IllegalArgumentException("Instances should be compatible.");
-        int k = a.size();
-        int denom = k * (k * k - 1);
+        long k = a.size();
+        long denom = k * (k * k - 1);
         double sum = 0.0;
         for (int i = 0; i < a.size(); i++) {
             double diff = (a.value(i) - b.value(i));
